+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # [unreleased]
 
+## Added
+* peek() function to Cache, to allow getting an item without requiring mutable access
+
 # 0.8.2
 
 ## Added